Add reference counting to Object<T> for safe memory mgmt
Introduce Reference<T> struct and ref counting in Object<T> to manage object lifetimes and memory more robustly. Update Object<T> to use Reference<T>* and implement copy, assignment, and destructor logic for reference management. Switch IDs and indexes to size_t for type safety. Refactor ObjectManager::remove for safer object removal. Update tests and clean up unused code.

+++ include/ObjectManager.hpp
@@ -22,8 +22,8 @@ namespace phun
 		ObjectManager& operator=(const ObjectManager&) = delete;
 		std::mutex mutex_;
 
-		std::vector<int>					objectIndexes_;
-		std::vector<int>							  ids_;
+		std::vector<size_t>					objectIndexes_;
+		std::vector<size_t>							  ids_;
 		std::vector<std::unique_ptr<ObjectBase>>  objects_;
 	};
 }
